web: organize time functions
