+++ web/src/time.test.ts
@@ -1,5 +1,5 @@
 import moment from "moment"
-import { formatBuildDuration } from "./format"
+import { formatBuildDuration } from "./time"
 
 it("format durations correctly", () => {
   function assertFormatted(expected: string, ms: number) {
